% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCsimtester.R
\docType{data}
\name{genetreeSample}
\alias{genetreeSample}
\title{Simulated gene tree dataset.}
\format{
A text file with 10,000 metric Newick gene trees on the taxa a,b,c,d,e
}
\description{
A dataset of 10,000 gene trees on 5 taxa simulated under the MSC on a species tree.
}
\details{
This simulated dataset was produced by SimPhy \insertRef{2016Mallo-simphy}{MSCsimtester},
using the species tree

((((a:10000,b:10000):10000,c:20000):10000,d:30000):10000,e:40000);

with population sizes

c(15000,25000,10000,1,1,1,1,1,12000)

and edges ordered by the \code{ape} function \code{read.tree}.

File is accessed as  \code{system.file("extdata","genetreeSample",package="MSCsimtester")}, for example
using the ape command:

\code{gts=read.tree(file=system.file("extdata","genetreeSample",package="MSCsimtester") )}
}
\references{
\insertRef{2016Mallo-simphy}{MSCsimtester}
}
