% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCsimtester.R
\name{print.ADtestOutput}
\alias{print.ADtestOutput}
\title{Print function for objects of class \code{ADtestOutput}.}
\usage{
\method{print}{ADtestOutput}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ADtestOutput}, as produced by \code{ADtest} function.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
NONE
}
\description{
Print function for objects of class \code{ADtestOutput}.
}
