% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCsimtester.R
\name{print.rootedTripleOutput}
\alias{print.rootedTripleOutput}
\title{Print function for objects of class \code{rootedTripleOutput}.}
\usage{
\method{print}{rootedTripleOutput}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rootedTripleOutput}, as produced by \code{rootedTriple} function.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
NONE
}
\description{
Print function for objects of class \code{rootedTripleOutput}.
}
\details{
Print function for objects of class \code{rootedTripleOutput}.
}
